Make first graphql call!
