+++ App/stores/util/index.ts
@@ -14,5 +14,6 @@
 // You should have received a copy of the GNU General Public License
 // along with Sh**t! I Smoke.  If not, see <http://www.gnu.org/licenses/>.
 
+export * from './createUser';
 export * from './fetchApi';
 export * from './fetchGpsPosition';
